+++ packages/react-fake-browser-ui/src/Button.ts
@@ -3,15 +3,15 @@ import styled from 'styled-components';
 export const Button = styled.button(({
   disabled,
 }) => ({
-  marginLeft: 5,
-  marginRight: 5,
+  marginLeft: '5px',
+  marginRight: '5px',
   padding: 0,
   border: 'none',
   backgroundColor: 'transparent',
   cursor: disabled ? 'default' : 'pointer',
-  width: 24,
-  height: 24,
-  borderRadius: 12,
+  width: '24px',
+  height: '24px',
+  borderRadius: '12px',
   outline: 'none',
   color: disabled ? '#999' : '#333',