+++ packages/react-fake-browser-ui/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.0 (12 jul 2023)
+
+### Improvement
+
+* Support `styled-components` version `6.x`
+
+### Breaking changes
+
+* `styled-components` should be added to dependencies manually
+
 ## 0.2.1 (29 oct 2022)
 
 ### Typescript