+++ packages/react-router-dom-fake-browser/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.0 (12 jul 2023)
+
+### Improvement
+
+* Support `styled-components` version `6.x`
+
+### Breaking changes
+
+* `styled-components` should be added to dependencies manually
+
 ## 0.3.0-alpha.1 (29 oct 2022)
 
 * Upgrade `react-router` to `v6`